Commonized msb() and same() functions.
